Разделил редьюсер полностью на части и применил combineReducers WIP
