+++ src/components/full-video-player/full-video-player.connect.jsx
@@ -1,6 +1,6 @@
 import FullVideoPlayer from './full-video-player';
 import {connect} from 'react-redux';
-import {ActionCreator} from '@reducer/reducer';
+import {ActionCreator} from '@reducer/player/player';
 import withFullVideoPlayer from '@hocs/with-full-video-player/with-full-video-player';
