+++ src/components/main/main.connect.jsx
@@ -1,7 +1,7 @@
 import Main from './main';
 import {connect} from 'react-redux';
 import {getFilmsByGenre} from './selectors/index';
-import {ActionCreator} from '@reducer/reducer';
+import {ActionCreator} from '@reducer/player/player';
 import withMain from '@hocs/with-main/with-main';
