+++ src/reducer/name-space.js
@@ -1,0 +1,10 @@
+const NameSpace = {
+  FILM: `FILM`,
+  PLAYER: `PLAYER`,
+  DATA: `DATA`,
+  FILTER: `FILTER`,
+  USER: `USER`,
+};
+
+
+export default NameSpace;
