+++ src/reducer/reducer.js
@@ -1,62 +1,16 @@
-import {extend} from '@utils/';
-import {FilterNames} from '@consts/';
+import {combineReducers} from 'redux';
+import {reducer as film} from './film/film';
+import {reducer as player} from './player/player';
+import {reducer as data} from './data/data';
+import {reducer as filter} from './filter/filter';
+import {reducer as user} from './user/user';
+import NameSpace from './name-space';
 
 
-const initialState = {
-  activeFilm: null,
-  films: [],
-  currentFilter: FilterNames.ALL,
-  activeFullVideoPlayer: null,
-};
-
-const ActionType = {
-  SET_ACTIVE_FILM: `SET_ACTIVE_FILM`,
-  CHANGE_FILTER: `CHANGE_FILTER`,
-  SET_ACTIVE_FULL_VIDEO_PLAYER: `SET_ACTIVE_FULL_VIDEO_PLAYER`,
-  REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
-};
-
-const ActionCreator = {
-  setActiveFilm: (film) => ({
-    type: ActionType.SET_ACTIVE_FILM,
-    payload: film,
-  }),
-  changeFilter: (currentFilter) => ({
-    type: ActionType.CHANGE_FILTER,
-    payload: currentFilter,
-  }),
-  setActiveFullVideoPlayer: (film) => ({
-    type: ActionType.SET_ACTIVE_FULL_VIDEO_PLAYER,
-    payload: film,
-  }),
-};
-
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ActionType.SET_ACTIVE_FILM:
-      return extend(state, {
-        activeFilm: action.payload,
-      });
-    case ActionType.CHANGE_FILTER:
-      return extend(state, {
-        currentFilter: action.payload,
-      });
-    case ActionType.SET_ACTIVE_FULL_VIDEO_PLAYER:
-      return extend(state, {
-        activeFullVideoPlayer: action.payload,
-      });
-    case ActionType.REQUIRED_AUTHORIZATION:
-      return Object.assign({}, state, {
-        authorizationStatus: action.payload,
-      });
-  }
-
-  return state;
-};
-
-
-export {
-  reducer,
-  ActionType,
-  ActionCreator
-};
+export default combineReducers({
+  [NameSpace.FILM]: film,
+  [NameSpace.PLAYER]: player,
+  [NameSpace.DATA]: data,
+  [NameSpace.FILTER]: filter,
+  [NameSpace.USER]: user,
+});
